[ADD] Test Code for Card Class
